# docs/contracts/phase1_2_openapi.yaml
openapi: 3.1.0
info:
  title: NoeticBraid Phase 1.2 API
  version: 1.2.0
  x-contract-version: 1.2.0
  x-status: AUTHORITATIVE
  x-frozen: true
  x-phase: "1.2"
  x-stage: "contract-1.2.0"
paths:
...
  /api/ledger/runs/aggregate:
    get:
      operationId: aggregate_run_record_api_ledger_runs_aggregate_get
      tags:
        - ledger
      summary: Aggregate run record
      parameters:
        - name: run_id
          in: query
          required: true
          schema:
            type: string
            pattern: "^run_[A-Za-z0-9_]+$"
            maxLength: 128
      responses:
        '200':
          description: Aggregated run record view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunRecordAggregate'
components:
  schemas:
...
    RunRecordAggregate:
      type: object
      additionalProperties: false
      required:
        - run_id
        - task_id
        - event_count
      properties:
        run_id:
          type: string
          pattern: "^run_[A-Za-z0-9_]+$"
          maxLength: 128
          description: Run identifier requested by the aggregate endpoint.
        task_id:
          type: string
          pattern: "^task_[A-Za-z0-9_]+$"
          maxLength: 128
          nullable: true
          description: Task identifier linked to the aggregate; nullable for empty streams.
        event_count:
          type: integer
          minimum: 0
          description: Computed count of RunRecord events in the aggregate input.
        first_event_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the first RunRecord event in the aggregate input.
        last_event_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last RunRecord event in the aggregate input.
        models_used:
          type: array
          nullable: true
          description: Union (dedup) of all model_refs across the run's events.
          items:
            type: string
            pattern: "^model_[A-Za-z0-9_]+$"
            maxLength: 128
        artifacts:
          type: array
          nullable: true
          description: Artifact references collected from artifact_created events.
          items:
            type: object
            additionalProperties: false
            required:
              - event_id
              - artifact_ref
            properties:
              event_id:
                type: string
              artifact_ref:
                type: string
        errors:
          type: array
          nullable: true
          description: Failure entries from task_failed and security_violation events; text fields follow nested NO-LEAK rules.
          items:
            type: object
            additionalProperties: false
            required:
              - event_id
            properties:
              event_id:
                type: string
              error_kind:
                type: string
                description: Best-effort classifier; unknown is allowed when the event payload lacks a kind.
              message:
                type: string
                description: "NO-LEAK: Derived only from RunRecord event payload and artifact_refs join; never from stdout, exception repr, environment, browser paths, or credential material."
        lessons_summary:
          type: array
          nullable: true
          description: Lesson entries from lesson_candidate_created events; text fields follow nested NO-LEAK rules.
          items:
            type: object
            additionalProperties: false
            required:
              - event_id
            properties:
              event_id:
                type: string
              lesson_candidate_text:
                type: string
                description: "NO-LEAK: Derived only from RunRecord event payload and artifact_refs join; never from stdout, exception repr, environment, browser paths, or credential material."